Fixing build issues
I noticed some issues with the vignettes, as well as rogue, unnecessary tidyverse functions in the Pareto optimality calculation and plotting. These have now been removed. Made changes to the references in the Pareto vignette.
I've taken the references to ape and phytools out of the Pareto vignette as well. Plotting landscape information on phylogenies can be it's own vignette perhaps?